edit .to_f

+++ idealweight.rb
@@ -2,6 +2,6 @@ puts "Укажите Ваше Имя:"
 name = gets.chomp
 
 puts "Укажите Ваш рост:"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 puts height <= 110 ? "#{name} Ваш вес уже оптимальный" : "#{name} Ваш оптимальный вес: #{(height - 110) * 1.15}"
